Backend/requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0

# Database dependencies
chromadb>=0.4.0
sqlalchemy>=2.0.0

# Web scraping dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.18.0
webdriver-manager>=4.0.0

# API and server dependencies
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Web search
tavily-python>=0.3.0

# Environment and configuration
python-dotenv>=1.0.0

# Additional utilities
aiohttp>=3.9.0 